\name{BEARscc-package}
\alias{BEARscc-package}
\alias{BEARscc}
\docType{package}
\title{
\packageTitle{BEARscc}
}
\description{
\packageDescription{BEARscc}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{BEARscc}
\packageIndices{BEARscc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BEARscc}

Maintainer: \packageMaintainer{BEARscc}
}
\references{
}

%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%}
